IIM Project code/requirements.txt
streamlit==1.24.1
openai>=0.27.0
sqlalchemy>=2.0
pydantic
PyPDF2
tqdm
python-dotenv